update wind scripts and filepaths

+++ vresc_0-prepare-shapefiles.py
@@ -1,5 +1,5 @@
 ###############
-### IMPORTS ###
+#%% IMPORTS ###
 import numpy as np
 import pandas as pd
 import os, sys, math
@@ -7,6 +7,7 @@ import geopandas as gpd
 import shapely
 import urllib.request, zipfile, tarfile, pickle
 from tqdm import tqdm, trange
+import h5pyd
 
 ### Only need zephyr for projpath
 import zephyr
@@ -16,7 +17,8 @@ datapath = zephyr.settings.datapath
 #################
 ### PROCEDURE ###
 
-###### Download shapefiles from external sources
+################################################
+#%%### Download shapefiles from external sources
 files = [
     ### (filepath, filename, url)
     (
@@ -38,7 +40,7 @@ files = [
         os.path.join(projpath,'in','Maps','HIFLD',''),
         'Political_Boundaries_(Area)-shp',
         ('https://opendata.arcgis.com/datasets/bee7adfd918e4393995f64e155a1bbdf_0.zip?'
-        'outSR=%7B%22wkid%22%3A102100%2C%22latestWkid%22%3A3857%7D')
+         'outSR=%7B%22wkid%22%3A102100%2C%22latestWkid%22%3A3857%7D')
     ),
     (
         os.path.join(projpath,'in','Maps','HIFLD',''),
@@ -81,9 +83,10 @@ for (filepath, filename, url) in files:
         zip_ref.close()
     print('Downloaded {}'.format(os.path.join(filepath,filename)))
 
-###### Make the US states zone map
+##################################
+#%%### Make the US states zone map
 usafile = os.path.join(projpath,'in','Maps','HIFLD','Political_Boundaries_(Area)-shp')
-dfzones = gpd.read_file(usafile)
+dfzones = gpd.read_file(usafile).to_crs({'init':'epsg:4326'})
 ### Downselect to states
 dfzones = dfzones.loc[
     (dfzones.COUNTRY=='USA')
@@ -98,15 +101,89 @@ dfzones = dfzones.loc[
 dfzones = dfzones.reset_index().rename(columns={'STATEABB':'state'})
 dfzones.state = dfzones.state.map(lambda x: x.replace('US-',''))
 ### Save it
-os.makedirs(os.path.join(projpath,'in','Maps','states'), exist_ok=True)
-dfzones.to_file(os.path.join(projpath,'in','Maps','states'))
+os.makedirs(os.path.join(projpath,'in','Maps','state'), exist_ok=True)
+dfzones.to_file(os.path.join(projpath,'in','Maps','state'))
 
-###### Prepare the national parks polygon to save time
-parks = gpd.read_file(datapath+'Maps/NPS/nps_boundary/')
-allparks = parks.loc[~parks.STATE.isin(['AK','HI','GU','MP','PR','AS','VI'])].copy()
-allparks['dummy'] = 0
-allparks = allparks.dissolve('dummy')
-polyallparks = allparks.loc[0, 'geometry']
+######################################################
+#%%### Prepare the national parks polygon to save time
+outfile = os.path.join(projpath,'in','Maps','NPS','nationalparks-all-poly.p')
+if not os.path.exists(outfile):
+    parks = gpd.read_file(datapath+'Maps/NPS/nps_boundary/')
+    allparks = parks.loc[~parks.STATE.isin(['AK','HI','GU','MP','PR','AS','VI'])].copy()
+    allparks['dummy'] = 0
+    allparks = allparks.dissolve('dummy')
+    polyallparks = allparks.loc[0, 'geometry']
+    ### Save it
+    with open(outfile, 'wb') as p:
+        pickle.dump(polyallparks, p)
+
+##############################################
+#%%### Create the list of WTK HSDS coordinates
+
+### Set up data access
+hsds = h5pyd.File("/nrel/wtk-us.h5", 'r')
+
+### Get the coordinates
+windcoords = hsds['coordinates']
+windcoords_data = windcoords[:][:]
+
+### Reshape and reformat
+nrows = windcoords_data.shape[0]
+ncols = windcoords_data.shape[1]
+
+rowsout_lat, rowsout_lon = {}, {}
+for row in trange(nrows):
+    rowsout_lat[row] = [x[0] for x in windcoords_data[row]]
+    rowsout_lon[row] = [x[1] for x in windcoords_data[row]]
+
+latsout = pd.DataFrame(rowsout_lat).T
+lonsout = pd.DataFrame(rowsout_lon).T
+
+latsout['row'] = latsout.index
+lonsout['row'] = lonsout.index
+
+dfout = (
+    pd.melt(latsout, var_name='col', value_name='latitude', id_vars='row')
+    .merge(
+        pd.melt(lonsout, var_name='col', value_name='longitude', id_vars='row'),
+        on=['row','col'])
+)
+
 ### Save it
-with open(os.path.join(projpath,'in','Maps','NPS','nationalparks-all-poly.p'), 'wb') as p:
-    pickle.dump(polyallparks, p)
+dfout.to_csv(
+    os.path.join(projpath,'io','geo','hsdscoords.gz'), 
+    compression='gzip', index=False)
+
+#######################################################################
+#%%### Create the list of WTK HSDS points within 0.5° of continental US
+### Get the USA states map
+usafile = os.path.join(projpath,'in','Maps','state')
+dfstates = gpd.read_file(usafile)
+### Dissolve to a single polygon
+dfstates['dummy'] = 0
+dfusa = dfstates.dissolve('dummy')
+dfusa.crs = {'init':'epsg:4326'}
+### Buffer it to 0.5°
+usa = dfusa.buffer(0.5).loc[0]
+
+### Load the WTK points
+dfcoords = pd.read_csv(
+    os.path.join(projpath,'io','geo','hsdscoords.gz'),
+    dtype={'row':int,'col':int,
+           'latitude':float,'longitude':float})
+
+### Get coordinates of downsampled WTK points
+wtkpoints = dfcoords.loc[(dfcoords.row % 2 == 0) & (dfcoords.col % 2 == 0)].copy()
+wtkpoints.rename(columns={'row':'row_full','col':'col_full'}, inplace=True)
+wtkpoints['row'] = (wtkpoints['row_full'] / 2).astype(int)
+wtkpoints['col'] = (wtkpoints['col_full'] / 2).astype(int)
+
+### Determine whether each point is inside buffered US polygon
+wtkpoints['usa'] = zephyr.toolbox.pointpolymap(
+    wtkpoints, usa, x='longitude',y='latitude',resetindex=False)
+
+### Save it
+wtkpoints.loc[wtkpoints.usa==True].drop('usa',axis=1).to_csv(
+    os.path.join(projpath,'io','geo','hsdscoords-usahifld-bufferhalfdeg.csv.gz'),
+    index=False, compression='gzip',
+)
